moved integration tests to separate files
